pcmcia(4): malo(4): tsleep(9) -> tsleep_nsec(9)
This command response sleep has no units, but in practice it is a 5
second timeout on 100hz platforms.  claudio@ notes that waiting 5
seconds for a command response from this hardware is excessive and
that the "500" was probably meant to be in milliseconds.  But absent a
problem he suggests leaving the sleep as-is, so we're keeping the 5
seconds.

ok claudio@