Do not reparent a traced child to ourself inside wait(2).
When a traced process _exit(2)s, its (tracing) parent tries to give it
back to the old parent.  In the case where the old parent is the same
as the tracing parent, there's no need to do this dance, so simply
remove it from the list of zombies and free its descriptors.

Fix a double report via wait(2) exposed by recent changes in make and
newly imported ptrace(2) regression from NetBSD.

Diagnosed with espie@ and guenther@, ok claudio@