Switch the legacy version to TLS1_2_VERSION when processing server hello.
Switch the legacy version when processing the server hello, rather than
when the client hello has been sent. This ensures that we accept a
response from the server that has a record version other than
TLS1_2_VERSION, as is the case where the server is negotiating TLSv1.0
or TLSv1.1.

Issue spotted by inoguchi@

ok inoguchi@ tb@