Revert previous
Total vcpu and memory are read from the PRI so constraint checks must not
be done before that, noted by kettenis.

The fact that "total_cpus" as a global variable (initialized with zero) is
always smaller than the total number of configured vcpus and the fact that
I only tested a negative example without a positive one made me jump the
trigger, sorry.