Move PowerBook5,4 audio from aoa(4) to snapper(4).
This adds the missing TAS3004 volume control.  Before, I put my ear
near the speaker (to hear audio), because I had no way to turn up the
volume.  Now, the default volume is much louder, so I use sndioctl(1)
to turn it down.

Check for the model string "PowerBook5,4".  This model's device tree
has compatible = "AOAKeylargo" for the audio, doesn't show the TAS3004
in the same way as other models.